requirements.txt
requests
python-dotenv
pandas
jinja2
schedule
